Install the `ruby-ntlm` gem to add NTLM support
In the same way it was done in the Cabildo Abierto project to
connect the SMTP client to the MS Exchange server using NTLM
authentication.

We did not cherry-picked the commit as it mixed up with other
changes. Also here we are fixing a version we know it works with
the Cabildo's current MS Exchange server.