Adapt spec to updated code
During the development of the related RP, the model user code was
updated:

-- Tenant.current_secrets.security[:lockable][:maximum_attempts]
++ Tenant.current_secrets.dig(:security, :lockable, :maximum_attempts)

But the "allow" was not adapted in the related test.

To make it possible for the test to pass again, we modify the allow according
to the modified code.

We are adding user model specs in order to test the maximum_attempts and
lock_in methods work for main tenant and custom tenants.