Show values at most logical place
Now, when showing `audited_changes` from an audit with the
`create` action we show the value in the `New value` column and the
`Old value` column it's empty. When the audit action is `destroy` we
show the value in the `Old value` column while the `New value`
column is empty. Before this change, we were using the `Old value`
always that was very confusing.